feat: Stripe webhook EP cleanup

+++ BlazorStripeExampleApp/Entities/Payment.cs
@@ -5,7 +5,7 @@ public class Payment
     public int Id { get; set; }
     public string SessionId { get; set; } = null!;
     public string CustomerEmail { get; set; } = null!;
-    public long AmountTotal { get; set; }
+    public decimal AmountTotal { get; set; }
     public string Currency { get; set; } = null!;
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
 }
